+++ app.js
@@ -1,10 +1,15 @@
-import { renderNode } from "./dom.service";
-import { virtualDOM } from "./domObjects";
+const express = require('express');
+const path = require('path');
 
-document.addEventListener("DOMContentLoaded", (event) => {
-    
-   const root = document.getElementById("root")
-   const renderedDom =  renderNode(virtualDOM)
-   root.appendChild(renderedDom)
+const app = express();
+const PORT = 3000;
 
+app.use(express.static(path.join(__dirname, 'public')));
+
+app.get('/', (_, res) => {
+  res.sendFile(path.join(__dirname, 'public', 'index.html'));
+});
+
+app.listen(PORT, () => {
+  console.log(`Server running at port:${PORT}`);
 });
