+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "dom-diff",
+  "version": "1.0.0",
+  "description": "",
+  "main": "app.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "nodemon app.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/rohankshah/simple-dom-diff.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "type": "commonjs",
+  "bugs": {
+    "url": "https://github.com/rohankshah/simple-dom-diff/issues"
+  },
+  "homepage": "https://github.com/rohankshah/simple-dom-diff#readme",
+  "dependencies": {
+    "express": "^5.1.0"
+  }
+}
