+++ public/index.html
@@ -5,13 +5,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>DOM Diff</title>
-    <script src="./app.js"></script>
+    <script type="module" src="./index.js"></script>
 </head>
 
-<body id="root">
-
-
-
-</body>
+<body id="root"></body>
 
 </html>
